move project resolver to its file
