+++ src/server-extension/resolvers/index.ts
@@ -1,1 +1,1 @@
-export { ProjectResolver } from "../resolver";
+export { ProjectResolver } from "../project-resolver";
